Restructured some files, Rebaked lights, Fixed lighting artifacts.
Countertops and random props were showing lighting artifacts which were ugly to look at. This was fixed by changing the lighting UV generation minimum scale. The models are default .01 scale and this was not set properly on the props. Fixing this got rid of the lightmap bleeding that was occurring on various items.